Use pytest for unit tests (#346)
* Use pytest for unit tests

Tests can be run with $ python3 -m pytest tests

I disabled slow tests by default, but they will be run by Jenkins and
can be run manually by adding the --runslow option

All tabular benchmarks have been moved to the same file as they all
depend on run_tabular_benchmarks. This is the simplest way to share
a helper function between pytest tests.

To run the tests in a specific test use

$ python3 -m pytest path_to_file:test_mytest

:test_mytest can be omitted to run all tests in a specific file.

* Use pytest for unit tests

Tests can be run with $ python3 -m pytest tests

I disabled slow tests by default, but they will be run by Jenkins and
can be run manually by adding the --runslow option

All tabular benchmarks have been moved to the same file as they all
depend on run_tabular_benchmarks. This is the simplest way to share
a helper function between pytest tests.

To run the tests in a specific test use

$ python3 -m pytest path_to_file:test_mytest

:test_mytest can be omitted to run all tests in a specific file.

* Install pytest in Jenkins

* Commit renamed check style file

* Add pytest to the requirements
